refactor(ui): extract BlessingTabRenderer.cs
