+++ PantheonWars/GUI/UI/MainDialogRenderer.cs
@@ -1,13 +1,10 @@
 using System;
-using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using ImGuiNET;
 using PantheonWars.GUI.State;
 using PantheonWars.GUI.UI.Components;
 using PantheonWars.GUI.UI.Renderers.Blessing;
 using PantheonWars.GUI.UI.Renderers.Civilization;
-using PantheonWars.GUI.UI.Renderers.Religion;
-using PantheonWars.Models;
 using Vintagestory.API.Client;
 
 namespace PantheonWars.GUI.UI;
@@ -111,86 +108,12 @@ internal static class MainDialogRenderer
                 manager.ReligionStateManager.DrawReligionTab(windowPos.X + x, windowPos.Y + y, width, contentHeight);
                 break;
             case MainDialogTab.Blessings: // Blessings
-                DrawBlessingsTab(manager, api, windowPos.X + x, windowPos.Y + y, width, contentHeight,
+                BlessingTabRenderer.DrawBlessingsTab(manager, api, windowPos.X + x, windowPos.Y + y, width, contentHeight,
                     windowWidth, windowHeight, deltaTime, onUnlockClicked, onCloseClicked);
                 break;
             case MainDialogTab.Civilization: // Civilization
                 CivilizationTabRenderer.Draw(manager, api, windowPos.X + x, windowPos.Y + y, width, contentHeight);
                 break;
-        }
-    }
-
-    private static void DrawBlessingsTab(
-        GuiDialogManager manager,
-        ICoreClientAPI api,
-        float x,
-        float y,
-        float width,
-        float height,
-        int windowWidth,
-        int windowHeight,
-        float deltaTime,
-        Action? onUnlockClicked,
-        Action? onCloseClicked)
-    {
-        const float infoPanelHeight = 200f;
-        const float padding = 16f;
-        const float actionButtonHeight = 36f;
-        const float actionButtonPadding = 16f;
-
-        // Track hovering state
-        string? hoveringBlessingId = null;
-
-        // Blessing Tree (split view)
-        var treeHeight = height - infoPanelHeight - padding;
-        BlessingTreeRenderer.Draw(
-            manager, api,
-            x, y, width, treeHeight,
-            deltaTime,
-            ref hoveringBlessingId
-        );
-
-        // Info Panel
-        var infoY = y + treeHeight + padding;
-        BlessingInfoRenderer.Draw(manager, api, x, infoY, width, infoPanelHeight);
-
-        // Action Buttons (overlay bottom-right)
-        var buttonY = windowHeight - actionButtonHeight - actionButtonPadding;
-        var buttonX = windowWidth - actionButtonPadding;
-        BlessingActionsRenderer.Draw(
-            manager, api,
-            ImGui.GetWindowPos().X + buttonX,
-            ImGui.GetWindowPos().Y + buttonY,
-            onUnlockClicked,
-            onCloseClicked
-        );
-
-        // Update manager hover state
-        manager.HoveringBlessingId = hoveringBlessingId;
-
-        // Tooltips
-        if (!string.IsNullOrEmpty(hoveringBlessingId))
-        {
-            var hoveringState = manager.ReligionStateManager.GetBlessingState(hoveringBlessingId);
-            if (hoveringState != null)
-            {
-                var allBlessings = new Dictionary<string, Blessing>();
-                foreach (var s in manager.ReligionStateManager.PlayerBlessingStates.Values)
-                    if (!allBlessings.ContainsKey(s.Blessing.BlessingId))
-                        allBlessings[s.Blessing.BlessingId] = s.Blessing;
-                foreach (var s in manager.ReligionStateManager.ReligionBlessingStates.Values)
-                    if (!allBlessings.ContainsKey(s.Blessing.BlessingId))
-                        allBlessings[s.Blessing.BlessingId] = s.Blessing;
-
-                var tooltipData = BlessingTooltipData.FromBlessingAndState(
-                    hoveringState.Blessing,
-                    hoveringState,
-                    allBlessings
-                );
-
-                var mousePos = ImGui.GetMousePos();
-                TooltipRenderer.Draw(tooltipData, mousePos.X, mousePos.Y, windowWidth, windowHeight);
-            }
         }
     }
 }
